requirements.txt
# Application dependencies. Dependency repository: https://pypi.org/
# gunicorn==20.0.4
# Flask==2.1.0
# pyjwt==2.1.0
# google-cloud-secret-manager==2.5.0
# requests==2.25.1
# google-cloud-dlp==0.13.0
# google-cloud-translate==3.2.1
# google-cloud-bigquery==2.3.1
# pytest==6.2.4
# pytest-mock==3.6.1
# google-auth

attrs==22.1.0
cachetools==4.2.4
certifi==2022.9.24
chardet==4.0.0
click==8.1.3
Flask==2.1.0
google-api-core==1.33.2
google-auth==1.35.0
google-cloud-bigquery==2.3.1
google-cloud-core==1.7.3
google-cloud-dlp==0.13.0
google-cloud-secret-manager==2.5.0
google-cloud-translate==3.2.1
google-crc32c==1.5.0
google-resumable-media==1.3.3
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==0.4.9
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pluggy==0.13.1
proto-plus==1.22.1
protobuf==3.20.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-mock==3.6.1
pytz==2022.6
PyYAML==6.0
requests==2.25.1
rsa==4.9
six==1.16.0
toml==0.10.2
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
slack_sdk
python-dotenv